Create Recon.sh
Subdomain Enumeration Script
This script automates the process of subdomain enumeration using three different tools: subfinder, sublist3r, and assetfinder. It takes a list of target domains from a file and saves the results for each target domain in a separate directory named after the domain.

Prerequisites
subfinder
sublist3r
assetfinder
These tools need to be installed and their paths need to be set in the script. In addition, the script uses other command-line tools like date and grep, which should be available in the system.

Usage
To use the script, follow these steps:

Make sure the prerequisites are installed.
Create a file named targets.txt in the same directory as the script.
Put the list of target domains, each on a separate line, in the targets.txt file.
Make the script executable by running chmod +x subenum.sh
Run the script by running ./subenum.sh
The script will create a directory for each target domain in the same directory as the script. The subdomains found for each target domain will be saved in separate files within the directory. The script will also create a file named subdomains.txt that contains all subdomains found for all target domains.

Output
The script will create the following output:

A directory for each target domain, named after the domain.
Files with the subdomains found for each tool used (subfinder, sublist3r, assetfinder), named after the tool and the date of the scan.
A file named subdomains.txt that contains all subdomains found for all target domains.
If the script is run multiple times, it will create a new file for each run of subdomains that were not previously found.
Modification
If you want to use other subdomain enumeration tools, you can modify the script to add them. Similarly, you can modify the output files to suit your needs. For example, you could change the name of the subdomains file or save the output in a different format. You could also modify the script to use different methods to notify you of the results, such as sending an email or posting to a Slack channel.
I removed telegram 🙈 I don't use it lol